Fix several bugs in Dockerfiles
We accidentally borked the CUDA install by chaging `CUDA_VERSION` to use
a minor version, which broke the symlink. This is now fixed.

Additionally, we fix some build errors in the `Dockerfile.torch` by
copying over the lines we dropped from the old file. Unfortunately I do
not have time to pare it down to just the essentials, but the full build
now works.